src/main/java/uk/gov/hmcts/reform/sandl/snlnotes/security/JwtTokenValidator.java
package uk.gov.hmcts.reform.sandl.snlnotes.security;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Slf4j
@Component
public class JwtTokenValidator {

    @Value("${jwt.secret}")
    private String secret;

    public String parseToken(String token) {
        Claims body = Jwts.parser()
                .setSigningKey(secret)
                .parseClaimsJws(token)
                .getBody();

        String username = body.getSubject();

        return username;
    }

}
